import-zips.py: add hack to help import ancient TeXShop releases
Richard provided me with lots of zip files with ancient TeXShop releases.
However, many of these .zip files were created freshly, in order to send them
to me. This results in the occasional file in the .zip having a brand new
date. Also, sometimes for other unknown reasons, a file here or there got a
newer change date.

This confuses the heuristic in import-zips.py, which takes the latest change
data of any file in the .zip as the date to assign the given TeXShop version
as release date. As a workaround, if most files in a .zip are ancient (date
before 2014 or so), we ignore the change dates of any files which were changed
2014 or later.